pyproject.toml
[tool.poetry]
name = "text-embeddings"
version = "0.1.0"
description = ""
authors = ["Ethan Thoma <ethanthoma@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python=">=3.9,<3.13"
pandas = "^2.1.1"
tenacity = "^8.2.3"
openai = "^0.28.1"
matplotlib = "^3.8.0"
plotly = "^5.17.0"
scipy = "^1.11.3"
scikit-learn = "^1.3.1"
tiktoken = "^0.5.1"
google-cloud-bigquery = "^3.12.0"
db-dtypes = "^1.1.1"
mypy = "^1.6.1"
pandas-stubs = "^2.1.1.230928"
numpy = "^1.26.1"
google-cloud-storage = "^2.12.0"
aiofiles = "^23.2.1"
ratemate = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

